requirements.txt
python-dotenv==1.0.0
Requests==2.31.0
SQLAlchemy==2.0.23
web3==6.11.3
